+++ packages/edit/src/components/extensions/image.ts
@@ -1,4 +1,3 @@
-import '@tiptap/extension-text-style';
 import Image from '@tiptap/extension-image';
 
 export default Image.configure({ inline: true }).extend({