+++ packages/display/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tailor-cms/ce-html-display
 
+## 0.2.1
+
+### Patch Changes
+
+- Fix migration issues with new TipTap version and cleanup code.
+
+## 0.2.0
+
+### Minor Changes
+
+- Bump Tiptap to v4 and migrate and cleanup text editor.
+
 ## 0.1.6
 
 ### Patch Changes